LibWeb: Bring HTMLHyperlinkElementUtils::follow_the_hyperlink to spec
This commit brings us brings us completely to spec on the link side
of things. It also fixes a bug wherein we would try make the browser
context navigate directly, even when it's not possible to do so
(for example, when the browsing content is within a different WebContent
process). This commit doesn't implement the remote browsing context
navigation part yet; that will come in a followup commit.